// src/types/models/activity-list-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Aftermeal API Docs
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface ActivityListResponseDto
 */
export interface ActivityListResponseDto {
    /**
     * 
     * @type {number}
     * @memberof ActivityListResponseDto
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof ActivityListResponseDto
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof ActivityListResponseDto
     */
    'location'?: string;
    /**
     * 
     * @type {number}
     * @memberof ActivityListResponseDto
     */
    'maxParticipants'?: number;
}